Move handler to dedicated folder and update with TypeScript static typing.
