--- handler.ts
@@ -1,16 +1,0 @@
-'use strict';
-
-module.exports.helloWorld = (event, context, callback) => {
-  const response = {
-    statusCode: 200,
-    headers: {
-      'Access-Control-Allow-Origin': '*', // Required for CORS support to work
-    },
-    body: JSON.stringify({
-      message: 'Hello World!',
-      input: event,
-    }),
-  };
-
-  callback(null, response);
-};
