+++ src/handlers/handler.js
@@ -1,0 +1,12 @@
+"use strict";
+exports.__esModule = true;
+exports.helloWorld = function (event, context, callback) {
+    var response = {
+        statusCode: 200,
+        body: JSON.stringify({
+            message: 'Hello World!',
+            input: event
+        })
+    };
+    callback(null, response);
+};
