--- src/hello_world.js
@@ -1,11 +1,0 @@
-"use strict";
-exports.__esModule = true;
-var http = require("http");
-var hostname = '127.0.0.1';
-var port = 3000;
-var server = http.createServer(function (req, res) {
-    res.statusCode = 200;
-    res.setHeader('Content-Type', 'text/plain');
-    res.end('Hello World\n');
-});
-server.listen(port, hostname);
