--- src/hello_world.ts
@@ -1,12 +1,0 @@
-import * as http from 'http';
-
-const hostname = '127.0.0.1';
-const port = 3000;
-
-const server = http.createServer((req: http.IncomingMessage, res: http.ServerResponse) => {
-  res.statusCode = 200;
-  res.setHeader('Content-Type', 'text/plain');
-  res.end('Hello World\n');
-});
-
-server.listen(port, hostname);
